+++ src/Plugin.FilePicker/Plugin.FilePicker/IFilePicker.cs
@@ -1,0 +1,18 @@
+﻿using System.Threading.Tasks;
+
+namespace Plugin.FilePicker
+{
+    /// <summary>
+    /// Interface for FilePicker
+    /// </summary>
+    public interface IFilePicker
+    {
+        Task<FileData> PickFile ();
+
+        Task<bool> SaveFile (FileData fileToSave);
+
+        void OpenFile (string fileToOpen);
+
+        void OpenFile (FileData fileToOpen);
+    }
+}
