+++ src/Plugin.FilePicker/Plugin.FilePicker/Platforms/monoandroid/FilePickerImplementation.cs
@@ -1,12 +1,4 @@
-using Android.App;
-using Android.Content;
-using Android.Runtime;
-using Java.IO;
-using Plugin.FilePicker.Abstractions;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Diagnostics;
+using Plugin.FilePicker.Platforms.monoandroid;
 
 namespace Plugin.FilePicker
 {
